requirements.txt
Flask==2.2.4
gunicorn==20.1.0
SQLAlchemy==1.4.49
psycopg2-binary==2.9.6
alembic==1.11.1
Flask-Migrate==4.0.4
requests==2.31.0
pytest==7.4.0
python-dotenv==1.0.0
cloud-sql-python-connector
pg8000